Refactor: Reorganizes files to make it easier to migrate to microservices on the future
#25
